convert quad => int64_t (from mycroft)
SCCS-vsn: 8.3

+++ sys/ufs/ffs/ffs_subr.c
@@ -30,7 +30,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)ffs_subr.c	8.2 (Berkeley) 09/21/93
+ *	@(#)ffs_subr.c	8.3 (Berkeley) 01/02/95
  */
 
 #include <sys/param.h>
@@ -89,7 +89,7 @@ void
 ffs_fragacct(fs, fragmap, fraglist, cnt)
 	struct fs *fs;
 	int fragmap;
-	long fraglist[];
+	int32_t fraglist[];
 	int cnt;
 {
 	int inblk;
